// src/gen-ts/src/models/CrupdateJob.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Company Management API
 * API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux, mouvements financiers, dépenses, achats et frais de transport. Basée sur le schéma de base de données fourni. 
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { JobStatus } from './JobStatus';
import {
    JobStatusFromJSON,
    JobStatusFromJSONTyped,
    JobStatusToJSON,
    JobStatusToJSONTyped,
} from './JobStatus';

/**
 * 
 * @export
 * @interface CrupdateJob
 */
export interface CrupdateJob {
    /**
     * 
     * @type {string}
     * @memberof CrupdateJob
     */
    comment?: string;
    /**
     * 
     * @type {string}
     * @memberof CrupdateJob
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CrupdateJob
     */
    company_id?: string;
    /**
     * 
     * @type {string}
     * @memberof CrupdateJob
     */
    description?: string;
    /**
     * 
     * @type {Date}
     * @memberof CrupdateJob
     */
    contract_signature_date?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CrupdateJob
     */
    start_date?: Date;
    /**
     * 
     * @type {Date}
     * @memberof CrupdateJob
     */
    end_date?: Date;
    /**
     * 
     * @type {JobStatus}
     * @memberof CrupdateJob
     */
    status?: JobStatus;
}



/**
 * Check if a given object implements the CrupdateJob interface.
 */
export function instanceOfCrupdateJob(value: object): value is CrupdateJob {
    return true;
}

export function CrupdateJobFromJSON(json: any): CrupdateJob {
    return CrupdateJobFromJSONTyped(json, false);
}

export function CrupdateJobFromJSONTyped(json: any, ignoreDiscriminator: boolean): CrupdateJob {
    if (json == null) {
        return json;
    }
    return {
        
        'comment': json['comment'] == null ? undefined : json['comment'],
        'id': json['id'] == null ? undefined : json['id'],
        'company_id': json['company_id'] == null ? undefined : json['company_id'],
        'description': json['description'] == null ? undefined : json['description'],
        'contract_signature_date': json['contract_signature_date'] == null ? undefined : (new Date(json['contract_signature_date'])),
        'start_date': json['start_date'] == null ? undefined : (new Date(json['start_date'])),
        'end_date': json['end_date'] == null ? undefined : (new Date(json['end_date'])),
        'status': json['status'] == null ? undefined : JobStatusFromJSON(json['status']),
    };
}

export function CrupdateJobToJSON(json: any): CrupdateJob {
    return CrupdateJobToJSONTyped(json, false);
}

export function CrupdateJobToJSONTyped(value?: CrupdateJob | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'comment': value['comment'],
        'id': value['id'],
        'company_id': value['company_id'],
        'description': value['description'],
        'contract_signature_date': value['contract_signature_date'] == null ? value['contract_signature_date'] : value['contract_signature_date'].toISOString().substring(0,10),
        'start_date': value['start_date'] == null ? value['start_date'] : value['start_date'].toISOString().substring(0,10),
        'end_date': value['end_date'] == null ? value['end_date'] : value['end_date'].toISOString().substring(0,10),
        'status': JobStatusToJSON(value['status']),
    };
}